main

+++ base/pipelines/fine_tuning.py
@@ -520,8 +520,7 @@ def train_lora_model(data, video_folder, args):
 
             assert text_features.dtype == last_hidden_state.dtype, "text_features and last_hidden_state must have the same dtype"
 
-            text_features = text_features.to(torch.float16)
-            last_hidden_state = last_hidden_state.to(torch.float16)
+            attention_layer = attention_layer.to(torch.float16)
 
             # Calcola l'attenzione
             attention_output, _ = attention_layer(text_features, last_hidden_state, last_hidden_state)
